requirements.txt
psutil
matplotlib
requests
pywin32
pystray
Pillow
